+++ src/main/resources/templates/administratorEditDriver.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Администратор: редактировать водителя</title>
+    <link rel="icon" href="/static/img/bus.png">
+</head>
+<body>
+<header th:insert="blocks/header.html :: header">
+</header>
+<main>
+    <div class="container">
+        <div class="row">
+            <div class="col s12" th:each="el : ${driver}">
+                <h5>Редактировать водителя:</h5>
+                <form th:action="'/administrations/administrator/drivers/' + ${el.id} + '/edit'" method="post">
+                    <input th:placeholder="${el.driverName}" type="text" name="driverName">
+                    <input th:placeholder="${el.driverSurname}" type="text" name="driverSurname">
+                    <input th:placeholder="${el.driverPhone}" type="text" name="driverPhone">
+                    <input type="hidden" name="_csrf" value="{{_csrf.token}}}">
+                    <button class="waves-effect waves-light btn" type="submit">Изменить водителя</button>
+                </form>
+            </div>
+        </div>
+    </div>
+</main>
+<script type="text/javascript" src="/static/js/materialize.min.js"></script>
+<footer th:insert="blocks/footer.html :: footer">
+</footer>
+</body>
+</html>
